Move redux action types back to store and import Loction type from React Router
